refactor(Add adapters for hand edits):
